tests/amdgpu: new test for Freesync-Video Mode
This tests transition between normal and FreeSync-Video modes and
measures the FPS to ensure vblank events are happening at the
expected rate. A SMPTE test pattern is displayed in the first mode,
and then a simple animation is displayed in the second mode, in
order to help the user to check if there are blanking with mode
transition.

v4: Rebase and refine commit message
v3: Remove conditional debugging calls and some comments.
v2: Rebase

Reviewed-by: Wayne Lin <Wayne.Lin@amd.com>
Signed-off-by: Solomon Chiu <solomon.chiu@amd.com>