tests/kms_psr2_sf: Add new testcases for moving plane without updates
Add new testcases which are moving cursor and overlay planes
without updating plane contents or setting any damage areas.

Cursor testcase is reproducing following bug:

https://gitlab.freedesktop.org/drm/intel/-/issues/5077

Current *-plane-move-sf-dmg-area is not suitable for this
purpose. It's blinking display on cleanup/prepare and also updating
contents of cursor plane. This makes it harder to observe possible
flickering. Patch is keeping original *-plane-move-sf-dmg-area as it
is and just adding a new testcase.

New testcase is moving the plane first inside pipe area. Then it's
moving it to x < 0 and last to y < 0. This is done step-by-step
because issue mentioned above is not triggered if the plane is just
moved to x < 0 or y < 0.

v2:
  - Keep PLANE_MOVE enum as it was
  - Remove changes to functions that are not really modified
  - Fix several copy paste errors
v3:
  - Use INTERACTIVE_PLANE_MOVE instead of PLANE_MOVE_NO_UPDATE
  - Fix coding style errors
  - Move psr_wait_entry before moving the plane
  - Cur_x/cur_y usage made more clear
v4:
  - Use PLANE_MOVE_CONTINUOUS instead of INTERACTIVE_PLANE_MOVE
  - Use 128 bytes for expected output string
  - Drop unnecessary comments

Cc: José Roberto de Souza <jose.souza@intel.com>
Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Jouni Högander <jouni.hogander@intel.com>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>