scripts/code_cov_gather*/sh: add help scripts for code coverage
When a Linux Kernel is built with gcov support, the compiler creates
some other object files (*.gcno) that are dependent of the compiler
version, and contain references to the source files.

At Kernel runtime at the test machine, counter files will be visible at
sysfs (*.gcda files), which are also on a compiler specific format.

In order to be able to properly parse the contents of the counters, the
information from 3 different places should be merged altogether:

	- Runtime counters: /sys/.../*.gcda and hiperlinks to *.gcno
	- Compile-time cross-reference object files: *.gcno
	- Kernel source code.

If the build machine is different than the source machine, some special
scripts are needed in order to either copy the source files to the
runtime machine or vice versa.

This is described at:

	https://www.kernel.org/doc/html/latest/dev-tools/gcov.html

Copy the two scripts described there as-is.

Further patches will modify them as needed.

Acked-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>