scripts/code_cov_capture.sh: add a script to use lcov on build+test machine
When the test machine is also the build machine, it is possible to run
lcov directly on it. That makes the IGT tests faster, as it won't need
to generate a results tarball, and the produced results will be
independent on the Kernel object files, making the output files more
portable.

It should be noticed that, in order to generate html files, the Kernel
source will still be needed, due to the detailed view.

Reviewed-by: Tomi Sarvela <tomi.p.sarvela@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>