Revert "tests/kms_async_flips: Make the crc test faster"
This reverts commit 9cb64a757d2ff1e180b1648e611439d94afd697d.

This patch was added to speed up the test and get better results with
shard-skls. But even with this we're still seeing failure on shard-skl.

Also, with this change we're seeing commit failure on setups having HDMI
displays connected. Although the failure is not caused by this patch,
reverting this patch anyway as we're not getting the intended improvement
on shard-skl as well.

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Ville Syrjälä <ville.syrjala@linux.intel.com>