runner: Override results for subtests even when there's dynamic subtests
igt_runner converts passing test results to warn if a test outputs
stuff to stderr. It also converts results if kernel logs contain
warnings to dmesg-warn or dmesg-fail, depending on the test's own
result.

Perform this same conversion to subtests even if they contain dynamic
subtests, now that we have the capability of keeping those around.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by:  Swati Sharma <swati2.sharma@intel.com>