tests/intel-ci/fast-feedback: Move gem_lmem_swapping
Since commit 37ec3b8ffd2c ("tests/gem_lmem_swapping: limit lmem to 4G")
we load the driver with limited 4GB of lmem for test execution. However
this means we also have to 1) reload the driver with that param set and
2) unload it after we are done. However tests/intel-ci/fast-feedback.testlist
has a dedicated section for tests that rely on module (un|re)loading so
not to impact other tests.

Move all gem_lmem_swapping to that section, after the coment:

	# All tests that do module unloading and reloading are executed last.
	# They will sometimes reveal issues of earlier tests leaving the
	# driver in a broken state that is not otherwise noticed in that test.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Acked-by: Petri Latvala <petri.latvala@intel.com>