tests/amdgpu/amd_psr: use PSR state enum replace the integer
[why & how]
The enum of AMDGPU PSR STATE may update along the time and should be
aligned to the code base of upstream. Instead of using hard-coded
integer, replace them with enum PSR_STATE would be preferred and we only
need to update the PSR_STATE enum if needed in the future.

Cc: Rodrigo Siqueira <rodrigo.siqueira@amd.com>
Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Leo Li <sunpeng.li@amd.com>
Cc: Jay Pillai <aurabindo.pillai@amd.com>
Cc: Wayne Lin <wayne.lin@amd.com>
Reviewed-by: Aurabindo Pillai <aurabindo.pillai@amd.com>
Signed-off-by: David Zhang <dingchen.zhang@amd.com>