tests/kms_plane_scaling: Fix modifier
The current test takes around 500.454s/connector/pipe on DG2
to complete. CI per test execution timeout is 120s.
In this test, we have fixed modifier as LINEAR and looped
through all the pixel formats which is the sole purpose of
this test. With this change, execution time has reduced to
340.297s.
To validate scaling with modifiers separate tests will be
introduced.

Cc: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Swati Sharma <swati2.sharma@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>