tests/i915_query: Test new DRM_I915_QUERY_GEOMETRY_SUBSLICES query
Ensure we get sensible responses for both valid and invalid usage of
this query item.

Note that unlike the traditional topology query we do not try to compare
the values returned against the old I915_PARAM ioctl.  Xe_HP already
uses a full 32-bit mask for subslices and we expect upcoming platforms
to increase the mask size to 64 or beyond (i.e., the hardware starts
using multiple registers to express the mask); since the old I915_PARAM
ioctl can only return a 32-bit value it will be unable to express the
full mask for upcoming platforms.

Cc: Matt Atwood <matthew.s.atwood@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>