tests/kms_atomic_transition: Separate commits for pipes with shared planes
Currently the helper method set_combinations() will set the properties
for each pipe/plane then push everything as a single commit. This will
cause issues for drivers which do not dedicate planes to pipes and allow
planes to be shared, as it causes the new pipe to overwrite the plane
properties that have been set by the old pipe. In addition, the DRM
framework forbids directly switching the assigned pipe for a plane within
the same commit [1].

To avoid this, we can first disable any pipes that currently are
handling a shared plane before setting the fb to be committed.

Note: This patch only fixes the issue for single display cases. Fixes
for multi-display cases will be addressed in a separate patch

Changes since V1:
- Removed shared plane bitmask
- Instead of immediately doing a NULL commit to disable the current pipe
  when primary plane is shared, disable the previous pipe using the
  primary plane before preparing the framebuffers for commit
- Removed commit for non-NULL framebuffers

[1]
https://gitlab.freedesktop.org/drm/msm/-/blob/msm-next/drivers/gpu/drm/drm_atomic.c#L695

Signed-off-by: Jessica Zhang <quic_jesszhan@quicinc.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>