code_cov_parse_info: fix error handling when opening files
While on bash scripts we use "||" for error handling, perl
accepts it, but it has a higher priority, thus would require
to use parenthesis before that.

That causes it to print a warning when the open fail while
reading filter files, instead of the original intent of
finishing and reporting an error.

So, let's use "or" instead.

Reviewed-by: Ch Sai Gowtham <sai.gowtham.ch@intel.com>
Reviewed-by: Andrzej Hajda <andrzej.hajda@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>