tests/gem_reset_stats: Test for shared reset domain
Added new subtest shared_reset_domain:
The test submits non-preemptible requests to all engines,
kills one and expects the rest to survive.
If one of those engines reset is RCS/CCS or multi-CCS
then expects only dependent engines to be reset.
It checks the status of context after reset.
Result:
1.If engine reset is one among dependent engines
then contexts of dependent engines to be victimized and
rest to be of noerror.
2.If engine reset is of non dependent engines then all
the contexts are of noerror.

Cc: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Signed-off-by: Priyanka Dandamudi <priyanka.dandamudi@intel.com>
Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>