tests/amdgpu/amd_psr: set visual-confirm default to disabled
[why & how]
For CI machine, by default it would directly run test case w/o any
manual confirmation. And visual confirm option is to validate the
PSR-SU feature as manual approach, which is for debugging purpose
and not for CI automated scheme.

Pre-set the "visual_confirm" to disable and only to enable the
visual confirm by manual.

changes in v2:
----------------------
- pre-set visual confirm option from enabled to disabled, and set
  it as required argument.

Cc: Rodrigo Siqueira <rodrigo.siqueira@amd.com>
Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Leo Li <sunpeng.li@amd.com>
Cc: Jay Pillai <aurabindo.pillai@amd.com>
Cc: Wayne Lin <wayne.lin@amd.com>

Signed-off-by: David Zhang <dingchen.zhang@amd.com>
Reviewed-by: Aurabindo Pillai <aurabindo.pillai@amd.com>