tests/amdgpu/amd_bypass: skip if DSC is enabled on the connector
[why & how]
Since the DSC is visually lossless but loss still exists after
compression/decompression at source/destination sides. So for any
bypass mode validation, which requires what received at sink is
consistent to what is transmitted at source, DSC is required to be
turned off.

skip the test run if the DSC is enabled on the connector (output).

Signed-off-by: David Zhang <dingchen.zhang@amd.com>
Acked-by: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>
Reviewed-by: Harry Wentland <harry.wentland@amd.com>