tests/kms_flip: Update logic to calculate test duration
Few of the subtests are intended to run only for one iteration with
duration of 0. But because of the existing logic to calculate duration,
even these subtests are running in loop for 0.5s. This is causing excessive
logging in the nonexisting-fb subtests.

Added check to use duration calculation logic only if duration is
non-zero in the first place.

Signed-off-by: Karthik B S <karthik.b.s@intel.com>
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>