tests/amdgpu/amd_psr: some fix for PSR-SU MPO test case
[why]
Analogous to usermode behavior, when there are primary plane and
overlay plane committed, alpha region of overlay corresponding to
the primary plane position/size should be set to zero to explicitly
display the content of primary plane (emulated as video playback in
this test case) in general.

The input arguments of width/height of the helper to paint the
framebuffer should not exceed the size of the created framebuffer.

[how]
- set the alpha region of overlay plane framebuffer to zero aligned
  to the position/size of the primary plane
- paint the primary plane framebuffer background/foreground w/ the
  valid width/height.

Cc: Rodrigo Siqueira <rodrigo.siqueira@amd.com>
Cc: Harry Wentland <harry.wentland@amd.com>
Cc: Leo Li <sunpeng.li@amd.com>
Cc: Jay Pillai <aurabindo.pillai@amd.com>
Cc: Wayne Lin <wayne.lin@amd.com>

changes in v2:
------------------
- correct the word "toplay" and replace w/ "overlay" to make it
  more clear and accurate

Signed-off-by: David Zhang <dingchen.zhang@amd.com>
Reviewed-by: Leo Li <sunpeng.li@amd.com>