tests/kms_writeback: Add helper method to detach pipe from output when test finishes
Currently, after a writeback test is run, the writeback connecter doesn't
release the pipe for other connecters to use. This can cause the device
to crash if another connector tries to use the same pipe (for other IGT
tests, for example).

In addition, a slight change [1] to DRM's implementation of the rmfb ioctl
means that calling igt_remove_fb() alone will not detach the CRTC --
instead, an explicit disable will be needed to do that.

[1]
https://gitlab.freedesktop.org/drm/msm/-/blob/msm-next/drivers/gpu/drm/drm_framebuffer.c#L997

Signed-off-by: Jessica Zhang <quic_jesszhan@quicinc.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>