tests/kms_concurrent: Skip the subtest if the resolution is not supported
The kms_concurrent is about testing mode setting with
reducing the resolution and then again increasing it,
for this the test will take the high resolution supported
by the connector and then calculate the lowest resolution,
if the calculated resolution is not supported by the connector
it will assign the default resolution of 1024x 768 without
checking.

Added check to skip the subtest if the current mode is smaller
than 1024 mode (igt_skip_on_f(mode_default->vdisplay < 1024))
becuase if the default resolution is greater then the highest
resolution supported by the connector then the test should skip.

Signed-off-by: Nidhi Gupta <nidhi1.gupta@intel.com>
Reviewed-by: Karthik B S <karthik.b.s@intel.com>