intel_gpu_top: Don't show client header if no kernel support
On kernels without support for the feature we should skip showing the
clients header to avoid confusing users.

Simply briefly open a render node to the selected device during init and
look if the relevant fields are present in the fdinfo data.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Issue: https://gitlab.freedesktop.org/drm/igt-gpu-tools/-/issues/120
Reviewed-by: Petri Latvala <petri.latvala@intel.com>