tests/i915: Use correct type for gen parameter
intel_gen is designed to return a big unsigned number for an unknown device.
The value is interpreted to be the newest generation possible, and the matching
configuration is used. Some functions incorrectly pass the gen value as a signed
number, meaning it becomes negative, and the oldest configuration is used
instead of the newest one. This leads to GPU hangs in some cases.

Update the definition of submit functions in gem_exec_gttfill and gem_softpin,
and has_cs_timestamp in gem_ctx_engines to use a correct type for gen parameter.

Signed-off-by: Karolina Drobnik <karolina.drobnik@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>